+++ src/views/components/PageSelection/PageSelection.tsx
@@ -7,11 +7,7 @@ import {
 import {
 	CustomPageSelection
 } from '@views/components/CustomPageSelection/CustomPageSelection.tsx'
-import style from '@views/components/GamePageList/GamePageList.module.scss'
-import { Dispatch, SetStateAction } from 'react'
-import {
-	IGamePageListViewModel
-} from '@src/view-models/interfaces/IGamePageListViewModel.ts'
+import { PageSwitch } from '@views/components/PageSwitch/PageSwitch.tsx'
 
 
 const viewModel = new PageSelectionViewModel()
@@ -26,37 +22,8 @@ export function PageSelection(props: IPageSelection) {
 			<h3>navigate by page</h3>
 
 			<PageSwitch pageIndices={pageIndices} {...props} />
-
-			<CustomPageSelection {...props}
-								 pageCount={pageIndices.length}/>
+			<CustomPageSelection {...props} pageCount={pageIndices.length}/>
 		</aside>
 	)
 }
 
-export interface IPageSwitch {
-	pageIndices: number[]
-	parentViewModel: IGamePageListViewModel
-	setCurrentPage: Dispatch<SetStateAction<number>>
-}
-
-function PageSwitch(props: IPageSwitch) {
-	const { pageIndices, parentViewModel, setCurrentPage } = props
-
-	const pageButtons = pageIndices.map(
-		pageIndex => (
-			<li key={`page-${pageIndex}`}>
-				<button data-page={pageIndex} type="button">
-					{pageIndex + 1}
-				</button>
-			</li>
-		))
-
-	return (
-		<menu className={style.PageSelection}
-			  onClick={(e) => {
-				  parentViewModel.changePage(e, setCurrentPage)
-			  }}>
-			{pageButtons}
-		</menu>
-	)
-}
